LPS-54424 Move generic code into superclass
The code needed to initialize both kinds of CMIS repositories is
identical, only changing the concrete instance that is initialized.

Sadly, most of the code in the subclasses is boilerplate, required to
inject OSGi managed dependencies.
